feat(ui): generate API types from OpenAPI
