+++ apps/ui/package.json
@@ -19,6 +19,8 @@
     "lint": "oxlint -c oxlint.json ./src",
     "format": "oxfmt --write ./src",
     "format:check": "oxfmt --check ./src",
+    "api-types:generate": "node scripts/generate-api-types.mjs",
+    "api-types:check": "node scripts/generate-api-types.mjs --check",
     "typecheck": "tsc --noEmit",
     "design:lint": "npx --yes @google/design.md@latest lint DESIGN.md",
     "test": "jest",
@@ -65,27 +67,11 @@
     "@types/react-dom": "^19",
     "jest": "^30.2.0",
     "jest-environment-jsdom": "^30.2.0",
+    "openapi-typescript": "^7.13.0",
     "oxfmt": "^0.51.0",
     "oxlint": "^1.49.0",
     "playwright": "^1.58.2",
     "tailwindcss": "^4.2.0",
     "typescript": "6.0.3"
-  },
-  "pnpm": {
-    "overrides": {
-      "@babel/core": "7.29.6",
-      "dompurify": "3.4.11",
-      "js-yaml": "4.2.0",
-      "postcss": "8.5.14",
-      "prismjs": "1.30.0",
-      "uuid": "11.1.1",
-      "ws": "^8.21.0"
-    },
-    "onlyBuiltDependencies": [
-      "@tailwindcss/oxide",
-      "esbuild",
-      "sharp",
-      "unrs-resolver"
-    ]
   }
 }
