+++ apps/ui/src/lib/api/schema-types.ts
@@ -1,0 +1,763 @@
+import type { components } from "./generated/openapi";
+
+type Schemas = components["schemas"];
+
+export type ApiSchema<Name extends keyof Schemas> = Schemas[Name];
+
+export type OpenApiActCompletedData = Schemas["ActCompletedData"];
+export type OpenApiActStartedData = Schemas["ActStartedData"];
+export type OpenApiAddA2aChannelHttpRequest = Schemas["AddA2aChannelHttpRequest"];
+export type OpenApiAddA2aChannelOutput = Schemas["AddA2aChannelOutput"];
+export type OpenApiAddApiEndpointChannelHttpRequest = Schemas["AddApiEndpointChannelHttpRequest"];
+export type OpenApiAddApiEndpointChannelOutput = Schemas["AddApiEndpointChannelOutput"];
+export type OpenApiAgent = Schemas["Agent"];
+export type OpenApiAgentAnalysisResponse = Schemas["AgentAnalysisResponse"];
+export type OpenApiAgentCapabilityConfig = Schemas["AgentCapabilityConfig"];
+export type OpenApiAgentMessage = Schemas["AgentMessage"];
+export type OpenApiAgentPreviewResponse = Schemas["AgentPreviewResponse"];
+export type OpenApiAgentStatus = Schemas["AgentStatus"];
+export type OpenApiAgentVersion = Schemas["AgentVersion"];
+export type OpenApiAgentVersionChangeKind = Schemas["AgentVersionChangeKind"];
+export type OpenApiAgentVersionDiffResponse = Schemas["AgentVersionDiffResponse"];
+export type OpenApiAgentVersionPolicy = Schemas["AgentVersionPolicy"];
+export type OpenApiAllowedAction = Schemas["AllowedAction"];
+export type OpenApiApp = Schemas["App"];
+export type OpenApiAppChannel = Schemas["AppChannel"];
+export type OpenApiAppEndpointAuthConfig = Schemas["AppEndpointAuthConfig"];
+export type OpenApiAppEndpointAuthMode = Schemas["AppEndpointAuthMode"];
+export type OpenApiAppEndpointAuthProviderConfig = Schemas["AppEndpointAuthProviderConfig"];
+export type OpenApiAppEndpointAuthRequirements = Schemas["AppEndpointAuthRequirements"];
+export type OpenApiAppRunBucket = Schemas["AppRunBucket"];
+export type OpenApiAppRunEvent = Schemas["AppRunEvent"];
+export type OpenApiAppRunListResponse = Schemas["AppRunListResponse"];
+export type OpenApiAppStatus = Schemas["AppStatus"];
+export type OpenApiBTreeMap = Schemas["BTreeMap"];
+export type OpenApiBackgroundProgress = Schemas["BackgroundProgress"];
+export type OpenApiBatchSetSecretsRequest = Schemas["BatchSetSecretsRequest"];
+export type OpenApiBatchSetSecretsResponse = Schemas["BatchSetSecretsResponse"];
+export type OpenApiBudgetEventData = Schemas["BudgetEventData"];
+export type OpenApiBuiltinTool = Schemas["BuiltinTool"];
+export type OpenApiCancelStatus = Schemas["CancelStatus"];
+export type OpenApiCancelTurnResponse = Schemas["CancelTurnResponse"];
+export type OpenApiCapabilityInfo = Schemas["CapabilityInfo"];
+export type OpenApiCapabilityLocalizationInfo = Schemas["CapabilityLocalizationInfo"];
+export type OpenApiCapabilityUsageData = Schemas["CapabilityUsageData"];
+export type OpenApiCapabilityUsageKind = Schemas["CapabilityUsageKind"];
+export type OpenApiCapabilityUsageRecord = Schemas["CapabilityUsageRecord"];
+export type OpenApiChannelType = Schemas["ChannelType"];
+export type OpenApiCheckAgentNameResponse = Schemas["CheckAgentNameResponse"];
+export type OpenApiCheckNameResponse = Schemas["CheckNameResponse"];
+export type OpenApiCircuitBreakerResponse = Schemas["CircuitBreakerResponse"];
+export type OpenApiCircuitBreakersListResponse = Schemas["CircuitBreakersListResponse"];
+export type OpenApiClientSideTool = Schemas["ClientSideTool"];
+export type OpenApiClientToolResult = Schemas["ClientToolResult"];
+export type OpenApiCommitRequest = Schemas["CommitRequest"];
+export type OpenApiCommitResult = Schemas["CommitResult"];
+export type OpenApiCompactionReason = Schemas["CompactionReason"];
+export type OpenApiCompactionStepData = Schemas["CompactionStepData"];
+export type OpenApiContentPart = Schemas["ContentPart"];
+export type OpenApiContextCompactedData = Schemas["ContextCompactedData"];
+export type OpenApiContextCompactingData = Schemas["ContextCompactingData"];
+export type OpenApiContextReportContribution = Schemas["ContextReportContribution"];
+export type OpenApiContextReportSection = Schemas["ContextReportSection"];
+export type OpenApiControls = Schemas["Controls"];
+export type OpenApiCopyFileRequest = Schemas["CopyFileRequest"];
+export type OpenApiCostTier = Schemas["CostTier"];
+export type OpenApiCreateAgentRequest = Schemas["CreateAgentRequest"];
+export type OpenApiCreateAgentVersionRequest = Schemas["CreateAgentVersionRequest"];
+export type OpenApiCreateAppRequest = Schemas["CreateAppRequest"];
+export type OpenApiCreateBranchRequest = Schemas["CreateBranchRequest"];
+export type OpenApiCreateDatabaseRequest = Schemas["CreateDatabaseRequest"];
+export type OpenApiCreateDeclarativeCapabilityRequest =
+  Schemas["CreateDeclarativeCapabilityRequest"];
+export type OpenApiCreateFileRequest = Schemas["CreateFileRequest"];
+export type OpenApiCreateHarnessRequest = Schemas["CreateHarnessRequest"];
+export type OpenApiCreateKnowledgeBaseRequest = Schemas["CreateKnowledgeBaseRequest"];
+export type OpenApiCreateKnowledgeEntryRequest = Schemas["CreateKnowledgeEntryRequest"];
+export type OpenApiCreateKnowledgeIndexRequest = Schemas["CreateKnowledgeIndexRequest"];
+export type OpenApiCreateMcpServerRequest = Schemas["CreateMcpServerRequest"];
+export type OpenApiCreateMemoryFileRequest = Schemas["CreateMemoryFileRequest"];
+export type OpenApiCreateMemoryRequest = Schemas["CreateMemoryRequest"];
+export type OpenApiCreateMemorySourceRequest = Schemas["CreateMemorySourceRequest"];
+export type OpenApiCreateMessageRequest = Schemas["CreateMessageRequest"];
+export type OpenApiCreateModelRequest = Schemas["CreateModelRequest"];
+export type OpenApiCreateOrganizationRequest = Schemas["CreateOrganizationRequest"];
+export type OpenApiCreatePaymentAccountRequest = Schemas["CreatePaymentAccountRequest"];
+export type OpenApiCreatePaymentPolicyRequest = Schemas["CreatePaymentPolicyRequest"];
+export type OpenApiCreateProviderRequest = Schemas["CreateProviderRequest"];
+export type OpenApiCreateSavedReportRequest = Schemas["CreateSavedReportRequest"];
+export type OpenApiCreateScheduleRequest = Schemas["CreateScheduleRequest"];
+export type OpenApiCreateSessionRequest = Schemas["CreateSessionRequest"];
+export type OpenApiCreateSkillRequest = Schemas["CreateSkillRequest"];
+export type OpenApiCreateTaskWebhookRequest = Schemas["CreateTaskWebhookRequest"];
+export type OpenApiCreateWorkspaceRequest = Schemas["CreateWorkspaceRequest"];
+export type OpenApiCredentialFormSchema = Schemas["CredentialFormSchema"];
+export type OpenApiDatabaseInfoResponse = Schemas["DatabaseInfoResponse"];
+export type OpenApiDatasetCatalog = Schemas["DatasetCatalog"];
+export type OpenApiDatasetCatalogEntry = Schemas["DatasetCatalogEntry"];
+export type OpenApiDatasetProjectorLag = Schemas["DatasetProjectorLag"];
+export type OpenApiDeclarativeCapability = Schemas["DeclarativeCapability"];
+export type OpenApiDeferrablePolicy = Schemas["DeferrablePolicy"];
+export type OpenApiDeleteAccountResponse = Schemas["DeleteAccountResponse"];
+export type OpenApiDeleteQuery = Schemas["DeleteQuery"];
+export type OpenApiDeleteResponse = Schemas["DeleteResponse"];
+export type OpenApiDiffQuery = Schemas["DiffQuery"];
+export type OpenApiDlqEntryResponse = Schemas["DlqEntryResponse"];
+export type OpenApiDlqListResponse = Schemas["DlqListResponse"];
+export type OpenApiDriverCredentialInfo = Schemas["DriverCredentialInfo"];
+export type OpenApiDriverId = Schemas["DriverId"];
+export type OpenApiEnqueueTaskOptions = Schemas["EnqueueTaskOptions"];
+export type OpenApiEnqueueTaskRequest = Schemas["EnqueueTaskRequest"];
+export type OpenApiEnqueueTaskResponse = Schemas["EnqueueTaskResponse"];
+export type OpenApiErrorResponse = Schemas["ErrorResponse"];
+export type OpenApiEvent = Schemas["Event"];
+export type OpenApiEventContext = Schemas["EventContext"];
+export type OpenApiEventData = Schemas["EventData"];
+export type OpenApiEventTypeCountOut = Schemas["EventTypeCountOut"];
+export type OpenApiEventsSummaryResult = Schemas["EventsSummaryResult"];
+export type OpenApiExecutionPhase = Schemas["ExecutionPhase"];
+export type OpenApiExportReportQueryRequest = Schemas["ExportReportQueryRequest"];
+export type OpenApiExportSavedReportRequest = Schemas["ExportSavedReportRequest"];
+export type OpenApiExportUserDataResponse = Schemas["ExportUserDataResponse"];
+export type OpenApiExportedOrganization = Schemas["ExportedOrganization"];
+export type OpenApiExportedPersonalAccessToken = Schemas["ExportedPersonalAccessToken"];
+export type OpenApiExportedUserProfile = Schemas["ExportedUserProfile"];
+export type OpenApiExternalActor = Schemas["ExternalActor"];
+export type OpenApiFailedReportingOutboxRow = Schemas["FailedReportingOutboxRow"];
+export type OpenApiFeatureFlagMap = Schemas["FeatureFlagMap"];
+export type OpenApiFieldType = Schemas["FieldType"];
+export type OpenApiFileInfo = Schemas["FileInfo"];
+export type OpenApiFileStat = Schemas["FileStat"];
+export type OpenApiFileWrittenData = Schemas["FileWrittenData"];
+export type OpenApiFinding = Schemas["Finding"];
+export type OpenApiFindingCategory = Schemas["FindingCategory"];
+export type OpenApiFindingLocation = Schemas["FindingLocation"];
+export type OpenApiFindingSeverity = Schemas["FindingSeverity"];
+export type OpenApiFindingSource = Schemas["FindingSource"];
+export type OpenApiForkAgentVersionRequest = Schemas["ForkAgentVersionRequest"];
+export type OpenApiForkSessionRequest = Schemas["ForkSessionRequest"];
+export type OpenApiFormField = Schemas["FormField"];
+export type OpenApiGetOrCreateChatSessionRequest = Schemas["GetOrCreateChatSessionRequest"];
+export type OpenApiGetQuery = Schemas["GetQuery"];
+export type OpenApiGetResponse = Schemas["GetResponse"];
+export type OpenApiGetSessionSandboxResponse = Schemas["GetSessionSandboxResponse"];
+export type OpenApiGitCommitInfo = Schemas["GitCommitInfo"];
+export type OpenApiGitDiff = Schemas["GitDiff"];
+export type OpenApiGitDiffEntry = Schemas["GitDiffEntry"];
+export type OpenApiGitDiffStats = Schemas["GitDiffStats"];
+export type OpenApiGitHubMemorySourceRequest = Schemas["GitHubMemorySourceRequest"];
+export type OpenApiGitHubMemorySourceResponse = Schemas["GitHubMemorySourceResponse"];
+export type OpenApiGitMemorySourceRequest = Schemas["GitMemorySourceRequest"];
+export type OpenApiGitMemorySourceResponse = Schemas["GitMemorySourceResponse"];
+export type OpenApiGitRefInfo = Schemas["GitRefInfo"];
+export type OpenApiGrepMatch = Schemas["GrepMatch"];
+export type OpenApiGrepRequest = Schemas["GrepRequest"];
+export type OpenApiGrepResult = Schemas["GrepResult"];
+export type OpenApiGrepResultEntry = Schemas["GrepResultEntry"];
+export type OpenApiGuardrailAction = Schemas["GuardrailAction"];
+export type OpenApiGuardrailExample = Schemas["GuardrailExample"];
+export type OpenApiGuardrailExamplesResponse = Schemas["GuardrailExamplesResponse"];
+export type OpenApiGuardrailStage = Schemas["GuardrailStage"];
+export type OpenApiGuardrailsDryRunHit = Schemas["GuardrailsDryRunHit"];
+export type OpenApiGuardrailsDryRunRequest = Schemas["GuardrailsDryRunRequest"];
+export type OpenApiGuardrailsDryRunResponse = Schemas["GuardrailsDryRunResponse"];
+export type OpenApiHarness = Schemas["Harness"];
+export type OpenApiHarnessExample = Schemas["HarnessExample"];
+export type OpenApiHarnessPreviewResponse = Schemas["HarnessPreviewResponse"];
+export type OpenApiHarnessStatus = Schemas["HarnessStatus"];
+export type OpenApiHealthCheckCaseResult = Schemas["HealthCheckCaseResult"];
+export type OpenApiHealthCheckRun = Schemas["HealthCheckRun"];
+export type OpenApiHealthCheckStatus = Schemas["HealthCheckStatus"];
+export type OpenApiHealthCheckSummary = Schemas["HealthCheckSummary"];
+export type OpenApiHealthResponse = Schemas["HealthResponse"];
+export type OpenApiImageContentPart = Schemas["ImageContentPart"];
+export type OpenApiImageFileContentPart = Schemas["ImageFileContentPart"];
+export type OpenApiImageInfo = Schemas["ImageInfo"];
+export type OpenApiImageUploadResponse = Schemas["ImageUploadResponse"];
+export type OpenApiImportOkfBundleRequest = Schemas["ImportOkfBundleRequest"];
+export type OpenApiInitialFile = Schemas["InitialFile"];
+export type OpenApiInputContentPart = Schemas["InputContentPart"];
+export type OpenApiInputMessage = Schemas["InputMessage"];
+export type OpenApiInputMessageData = Schemas["InputMessageData"];
+export type OpenApiInvocationSessionMode = Schemas["InvocationSessionMode"];
+export type OpenApiKeyValueInfo = Schemas["KeyValueInfo"];
+export type OpenApiKnowledgeBaseResponse = Schemas["KnowledgeBaseResponse"];
+export type OpenApiKnowledgeEntryResponse = Schemas["KnowledgeEntryResponse"];
+export type OpenApiKnowledgeIndexDocumentResponse = Schemas["KnowledgeIndexDocumentResponse"];
+export type OpenApiKnowledgeIndexResponse = Schemas["KnowledgeIndexResponse"];
+export type OpenApiLatestHealthCheckRun = Schemas["LatestHealthCheckRun"];
+export type OpenApiLeasedResource = Schemas["LeasedResource"];
+export type OpenApiLeasedResourceStatus = Schemas["LeasedResourceStatus"];
+export type OpenApiListAppRunsQuery = Schemas["ListAppRunsQuery"];
+export type OpenApiListExecutionsQuery = Schemas["ListExecutionsQuery"];
+export type OpenApiListKnowledgeBasesQuery = Schemas["ListKnowledgeBasesQuery"];
+export type OpenApiListKnowledgeEntriesQuery = Schemas["ListKnowledgeEntriesQuery"];
+export type OpenApiListKnowledgeIndexesQuery = Schemas["ListKnowledgeIndexesQuery"];
+export type OpenApiListMemoriesQuery = Schemas["ListMemoriesQuery"];
+export type OpenApiListResponse_Agent = Schemas["ListResponse_Agent"];
+export type OpenApiListResponse_CapabilityInfo = Schemas["ListResponse_CapabilityInfo"];
+export type OpenApiListResponse_DatabaseInfoResponse = Schemas["ListResponse_DatabaseInfoResponse"];
+export type OpenApiListResponse_Event = Schemas["ListResponse_Event"];
+export type OpenApiListResponse_FileInfo = Schemas["ListResponse_FileInfo"];
+export type OpenApiListResponse_GrepResult = Schemas["ListResponse_GrepResult"];
+export type OpenApiListResponse_GrepResultEntry = Schemas["ListResponse_GrepResultEntry"];
+export type OpenApiListResponse_Harness = Schemas["ListResponse_Harness"];
+export type OpenApiListResponse_KeyValueInfo = Schemas["ListResponse_KeyValueInfo"];
+export type OpenApiListResponse_KnowledgeBaseResponse =
+  Schemas["ListResponse_KnowledgeBaseResponse"];
+export type OpenApiListResponse_KnowledgeEntryResponse =
+  Schemas["ListResponse_KnowledgeEntryResponse"];
+export type OpenApiListResponse_KnowledgeIndexDocumentResponse =
+  Schemas["ListResponse_KnowledgeIndexDocumentResponse"];
+export type OpenApiListResponse_KnowledgeIndexResponse =
+  Schemas["ListResponse_KnowledgeIndexResponse"];
+export type OpenApiListResponse_McpServer = Schemas["ListResponse_McpServer"];
+export type OpenApiListResponse_MemoryFileInfo = Schemas["ListResponse_MemoryFileInfo"];
+export type OpenApiListResponse_MemoryResponse = Schemas["ListResponse_MemoryResponse"];
+export type OpenApiListResponse_Message = Schemas["ListResponse_Message"];
+export type OpenApiListResponse_OrganizationResponse = Schemas["ListResponse_OrganizationResponse"];
+export type OpenApiListResponse_SavedReport = Schemas["ListResponse_SavedReport"];
+export type OpenApiListResponse_SecretInfo = Schemas["ListResponse_SecretInfo"];
+export type OpenApiListResponse_Skill = Schemas["ListResponse_Skill"];
+export type OpenApiListResponse_TaskWebhookResponse = Schemas["ListResponse_TaskWebhookResponse"];
+export type OpenApiListResponse_User = Schemas["ListResponse_User"];
+export type OpenApiListResponse_WithUrls_App = Schemas["ListResponse_WithUrls_App"];
+export type OpenApiListResponse_WithUrls_DeclarativeCapability =
+  Schemas["ListResponse_WithUrls_DeclarativeCapability"];
+export type OpenApiListResponse_WithUrls_McpServer = Schemas["ListResponse_WithUrls_McpServer"];
+export type OpenApiListResponse_WithUrls_Model = Schemas["ListResponse_WithUrls_Model"];
+export type OpenApiListResponse_WithUrls_ModelWithProvider =
+  Schemas["ListResponse_WithUrls_ModelWithProvider"];
+export type OpenApiListResponse_WithUrls_Provider = Schemas["ListResponse_WithUrls_Provider"];
+export type OpenApiListResponse_WithUrls_ResourceWithCounts_Harness =
+  Schemas["ListResponse_WithUrls_ResourceWithCounts_Harness"];
+export type OpenApiListResponse_WithUrls_Skill = Schemas["ListResponse_WithUrls_Skill"];
+export type OpenApiListResponse_WorkspaceResponse = Schemas["ListResponse_WorkspaceResponse"];
+export type OpenApiListSchedulesQuery = Schemas["ListSchedulesQuery"];
+export type OpenApiListUsersQuery = Schemas["ListUsersQuery"];
+export type OpenApiListWorkspacesQuery = Schemas["ListWorkspacesQuery"];
+export type OpenApiLlmCompactionInfo = Schemas["LlmCompactionInfo"];
+export type OpenApiLlmGenerationData = Schemas["LlmGenerationData"];
+export type OpenApiLlmGenerationMetadata = Schemas["LlmGenerationMetadata"];
+export type OpenApiLlmGenerationOutput = Schemas["LlmGenerationOutput"];
+export type OpenApiLlmPromptCacheInfo = Schemas["LlmPromptCacheInfo"];
+export type OpenApiLlmRequestOptions = Schemas["LlmRequestOptions"];
+export type OpenApiLlmRetryInfo = Schemas["LlmRetryInfo"];
+export type OpenApiLlmToolSearchInfo = Schemas["LlmToolSearchInfo"];
+export type OpenApiLogQuery = Schemas["LogQuery"];
+export type OpenApiManageSessionSandboxRequest = Schemas["ManageSessionSandboxRequest"];
+export type OpenApiManageSessionSandboxResponse = Schemas["ManageSessionSandboxResponse"];
+export type OpenApiManualMemorySourceResponse = Schemas["ManualMemorySourceResponse"];
+export type OpenApiMcpErrorCategory = Schemas["McpErrorCategory"];
+export type OpenApiMcpErrorCode = Schemas["McpErrorCode"];
+export type OpenApiMcpExecuteError = Schemas["McpExecuteError"];
+export type OpenApiMcpServer = Schemas["McpServer"];
+export type OpenApiMcpServerAuthMode = Schemas["McpServerAuthMode"];
+export type OpenApiMcpServerStatus = Schemas["McpServerStatus"];
+export type OpenApiMcpServerTransportType = Schemas["McpServerTransportType"];
+export type OpenApiMcpToolAnnotations = Schemas["McpToolAnnotations"];
+export type OpenApiMemoryFile = Schemas["MemoryFile"];
+export type OpenApiMemoryFileInfo = Schemas["MemoryFileInfo"];
+export type OpenApiMemoryResponse = Schemas["MemoryResponse"];
+export type OpenApiMemorySourceResponse = Schemas["MemorySourceResponse"];
+export type OpenApiMessage = Schemas["Message"];
+export type OpenApiMessageBody = Schemas["MessageBody"];
+export type OpenApiMessageRole = Schemas["MessageRole"];
+export type OpenApiMetricsPoint = Schemas["MetricsPoint"];
+export type OpenApiMetricsTimeSeriesResponse = Schemas["MetricsTimeSeriesResponse"];
+export type OpenApiModality = Schemas["Modality"];
+export type OpenApiModel = Schemas["Model"];
+export type OpenApiModelCost = Schemas["ModelCost"];
+export type OpenApiModelLimits = Schemas["ModelLimits"];
+export type OpenApiModelMetadata = Schemas["ModelMetadata"];
+export type OpenApiModelModalities = Schemas["ModelModalities"];
+export type OpenApiModelProfile = Schemas["ModelProfile"];
+export type OpenApiModelSource = Schemas["ModelSource"];
+export type OpenApiModelVendor = Schemas["ModelVendor"];
+export type OpenApiModelWithProvider = Schemas["ModelWithProvider"];
+export type OpenApiMoveFileRequest = Schemas["MoveFileRequest"];
+export type OpenApiNetworkAccessList = Schemas["NetworkAccessList"];
+export type OpenApiOkfFileInput = Schemas["OkfFileInput"];
+export type OpenApiOkfImportSummary = Schemas["OkfImportSummary"];
+export type OpenApiOrgFeatureFlagSetting = Schemas["OrgFeatureFlagSetting"];
+export type OpenApiOrgFeatureFlagsSettingsResponse = Schemas["OrgFeatureFlagsSettingsResponse"];
+export type OpenApiOrganizationResponse = Schemas["OrganizationResponse"];
+export type OpenApiOutputMessageCompletedData = Schemas["OutputMessageCompletedData"];
+export type OpenApiOutputMessageDeltaData = Schemas["OutputMessageDeltaData"];
+export type OpenApiOutputMessageReplacedData = Schemas["OutputMessageReplacedData"];
+export type OpenApiOutputMessageStartedData = Schemas["OutputMessageStartedData"];
+export type OpenApiPaginatedResponse_Session = Schemas["PaginatedResponse_Session"];
+export type OpenApiPaginatedResponse_WithUrls_CapabilityInfo =
+  Schemas["PaginatedResponse_WithUrls_CapabilityInfo"];
+export type OpenApiPaginatedResponse_WithUrls_ResourceWithCounts_Agent =
+  Schemas["PaginatedResponse_WithUrls_ResourceWithCounts_Agent"];
+export type OpenApiPaginatedResponse_WithUrls_Session =
+  Schemas["PaginatedResponse_WithUrls_Session"];
+export type OpenApiPaymentAccount = Schemas["PaymentAccount"];
+export type OpenApiPaymentAttempt = Schemas["PaymentAttempt"];
+export type OpenApiPaymentOwnerType = Schemas["PaymentOwnerType"];
+export type OpenApiPaymentPolicy = Schemas["PaymentPolicy"];
+export type OpenApiPaymentRail = Schemas["PaymentRail"];
+export type OpenApiPaymentStatus = Schemas["PaymentStatus"];
+export type OpenApiPostTaskMessageBody = Schemas["PostTaskMessageBody"];
+export type OpenApiPreviewAgentRequest = Schemas["PreviewAgentRequest"];
+export type OpenApiPreviewHarnessRequest = Schemas["PreviewHarnessRequest"];
+export type OpenApiPrincipalKind = Schemas["PrincipalKind"];
+export type OpenApiPrincipalSummary = Schemas["PrincipalSummary"];
+export type OpenApiProfileResponse = Schemas["ProfileResponse"];
+export type OpenApiProjectorRunQuery = Schemas["ProjectorRunQuery"];
+export type OpenApiProjectorRunResult = Schemas["ProjectorRunResult"];
+export type OpenApiPromptCacheStrategy = Schemas["PromptCacheStrategy"];
+export type OpenApiProvider = Schemas["Provider"];
+export type OpenApiProviderStatus = Schemas["ProviderStatus"];
+export type OpenApiProviderTraceConfig = Schemas["ProviderTraceConfig"];
+export type OpenApiProvidersConfigResponse = Schemas["ProvidersConfigResponse"];
+export type OpenApiReasonCompletedData = Schemas["ReasonCompletedData"];
+export type OpenApiReasonItemData = Schemas["ReasonItemData"];
+export type OpenApiReasonRecoveredData = Schemas["ReasonRecoveredData"];
+export type OpenApiReasonStartedData = Schemas["ReasonStartedData"];
+export type OpenApiReasonThinkingCompletedData = Schemas["ReasonThinkingCompletedData"];
+export type OpenApiReasonThinkingDeltaData = Schemas["ReasonThinkingDeltaData"];
+export type OpenApiReasonThinkingStartedData = Schemas["ReasonThinkingStartedData"];
+export type OpenApiReasoningConfig = Schemas["ReasoningConfig"];
+export type OpenApiReasoningEffort = Schemas["ReasoningEffort"];
+export type OpenApiReasoningEffortConfig = Schemas["ReasoningEffortConfig"];
+export type OpenApiReasoningEffortValue = Schemas["ReasoningEffortValue"];
+export type OpenApiRecoveryMode = Schemas["RecoveryMode"];
+export type OpenApiRegenerateA2aApiKeyOutput = Schemas["RegenerateA2aApiKeyOutput"];
+export type OpenApiRegenerateApiEndpointApiKeyOutput = Schemas["RegenerateApiEndpointApiKeyOutput"];
+export type OpenApiReportColumn = Schemas["ReportColumn"];
+export type OpenApiReportColumnKind = Schemas["ReportColumnKind"];
+export type OpenApiReportExport = Schemas["ReportExport"];
+export type OpenApiReportExportFormat = Schemas["ReportExportFormat"];
+export type OpenApiReportFilter = Schemas["ReportFilter"];
+export type OpenApiReportFilterOp = Schemas["ReportFilterOp"];
+export type OpenApiReportOrderBy = Schemas["ReportOrderBy"];
+export type OpenApiReportOrderDirection = Schemas["ReportOrderDirection"];
+export type OpenApiReportQuery = Schemas["ReportQuery"];
+export type OpenApiReportResult = Schemas["ReportResult"];
+export type OpenApiReportTimeRange = Schemas["ReportTimeRange"];
+export type OpenApiReportingBackfillRequest = Schemas["ReportingBackfillRequest"];
+export type OpenApiReportingBackfillResult = Schemas["ReportingBackfillResult"];
+export type OpenApiReportingDiagnostics = Schemas["ReportingDiagnostics"];
+export type OpenApiReportingOutboxDiagnostics = Schemas["ReportingOutboxDiagnostics"];
+export type OpenApiResolveOrgResponse = Schemas["ResolveOrgResponse"];
+export type OpenApiResourceConfigResponse = Schemas["ResourceConfigResponse"];
+export type OpenApiResourceStatsResponse = Schemas["ResourceStatsResponse"];
+export type OpenApiRiskLevel = Schemas["RiskLevel"];
+export type OpenApiRollbackAgentVersionRequest = Schemas["RollbackAgentVersionRequest"];
+export type OpenApiSavedReport = Schemas["SavedReport"];
+export type OpenApiSavedReportDashboardMetadata = Schemas["SavedReportDashboardMetadata"];
+export type OpenApiScheduleExecutionResponse = Schemas["ScheduleExecutionResponse"];
+export type OpenApiScheduleExecutionsListResponse = Schemas["ScheduleExecutionsListResponse"];
+export type OpenApiScheduleResponse = Schemas["ScheduleResponse"];
+export type OpenApiScheduleStatsResponse = Schemas["ScheduleStatsResponse"];
+export type OpenApiScheduleTarget = Schemas["ScheduleTarget"];
+export type OpenApiScheduleTargetResponse = Schemas["ScheduleTargetResponse"];
+export type OpenApiSchedulesListResponse = Schemas["SchedulesListResponse"];
+export type OpenApiSchemaResponse = Schemas["SchemaResponse"];
+export type OpenApiSecretInfo = Schemas["SecretInfo"];
+export type OpenApiSendSignalRequest = Schemas["SendSignalRequest"];
+export type OpenApiSession = Schemas["Session"];
+export type OpenApiSessionActivatedData = Schemas["SessionActivatedData"];
+export type OpenApiSessionContextReport = Schemas["SessionContextReport"];
+export type OpenApiSessionFile = Schemas["SessionFile"];
+export type OpenApiSessionIdledData = Schemas["SessionIdledData"];
+export type OpenApiSessionRef = Schemas["SessionRef"];
+export type OpenApiSessionResourceEntry = Schemas["SessionResourceEntry"];
+export type OpenApiSessionResourceStatus = Schemas["SessionResourceStatus"];
+export type OpenApiSessionSandboxAction = Schemas["SessionSandboxAction"];
+export type OpenApiSessionSandboxStatusValue = Schemas["SessionSandboxStatusValue"];
+export type OpenApiSessionStartedData = Schemas["SessionStartedData"];
+export type OpenApiSessionStatsResponse = Schemas["SessionStatsResponse"];
+export type OpenApiSessionStatus = Schemas["SessionStatus"];
+export type OpenApiSessionTask = Schemas["SessionTask"];
+export type OpenApiSessionTaskDetail = Schemas["SessionTaskDetail"];
+export type OpenApiSessionTaskEventData = Schemas["SessionTaskEventData"];
+export type OpenApiSessionTaskState = Schemas["SessionTaskState"];
+export type OpenApiSetDefaultAgentVersionRequest = Schemas["SetDefaultAgentVersionRequest"];
+export type OpenApiSideEffectClass = Schemas["SideEffectClass"];
+export type OpenApiSkill = Schemas["Skill"];
+export type OpenApiSkillContent = Schemas["SkillContent"];
+export type OpenApiSkillFileEntry = Schemas["SkillFileEntry"];
+export type OpenApiSkillSourceType = Schemas["SkillSourceType"];
+export type OpenApiSkillStatus = Schemas["SkillStatus"];
+export type OpenApiSkillUsage = Schemas["SkillUsage"];
+export type OpenApiSkillValidationResult = Schemas["SkillValidationResult"];
+export type OpenApiStatRequest = Schemas["StatRequest"];
+export type OpenApiSubmitToolResultsRequest = Schemas["SubmitToolResultsRequest"];
+export type OpenApiSubmitToolResultsResponse = Schemas["SubmitToolResultsResponse"];
+export type OpenApiSuccessResponse = Schemas["SuccessResponse"];
+export type OpenApiSwitchOrgRequest = Schemas["SwitchOrgRequest"];
+export type OpenApiSwitchOrgResponse = Schemas["SwitchOrgResponse"];
+export type OpenApiSyncModelsResponse = Schemas["SyncModelsResponse"];
+export type OpenApiTaskArtifact = Schemas["TaskArtifact"];
+export type OpenApiTaskError = Schemas["TaskError"];
+export type OpenApiTaskInputRequest = Schemas["TaskInputRequest"];
+export type OpenApiTaskLinks = Schemas["TaskLinks"];
+export type OpenApiTaskMessage = Schemas["TaskMessage"];
+export type OpenApiTaskMessageDirection = Schemas["TaskMessageDirection"];
+export type OpenApiTaskMessageEventData = Schemas["TaskMessageEventData"];
+export type OpenApiTaskMessagePart = Schemas["TaskMessagePart"];
+export type OpenApiTaskResponse = Schemas["TaskResponse"];
+export type OpenApiTaskWakePolicy = Schemas["TaskWakePolicy"];
+export type OpenApiTaskWebhookResponse = Schemas["TaskWebhookResponse"];
+export type OpenApiTasksListResponse = Schemas["TasksListResponse"];
+export type OpenApiTextContentPart = Schemas["TextContentPart"];
+export type OpenApiTokenUsage = Schemas["TokenUsage"];
+export type OpenApiToolCall = Schemas["ToolCall"];
+export type OpenApiToolCallContentPart = Schemas["ToolCallContentPart"];
+export type OpenApiToolCallRepairedData = Schemas["ToolCallRepairedData"];
+export type OpenApiToolCallRequestedData = Schemas["ToolCallRequestedData"];
+export type OpenApiToolCallSummary = Schemas["ToolCallSummary"];
+export type OpenApiToolCompletedData = Schemas["ToolCompletedData"];
+export type OpenApiToolDefinition = Schemas["ToolDefinition"];
+export type OpenApiToolDefinitionSummary = Schemas["ToolDefinitionSummary"];
+export type OpenApiToolHints = Schemas["ToolHints"];
+export type OpenApiToolOutputDeltaData = Schemas["ToolOutputDeltaData"];
+export type OpenApiToolPolicy = Schemas["ToolPolicy"];
+export type OpenApiToolProgressData = Schemas["ToolProgressData"];
+export type OpenApiToolResultContentPart = Schemas["ToolResultContentPart"];
+export type OpenApiToolStartedData = Schemas["ToolStartedData"];
+export type OpenApiTranscriptRepairAction = Schemas["TranscriptRepairAction"];
+export type OpenApiTranscriptRepairedData = Schemas["TranscriptRepairedData"];
+export type OpenApiTriggerResponse = Schemas["TriggerResponse"];
+export type OpenApiTurnCancelledData = Schemas["TurnCancelledData"];
+export type OpenApiTurnCompletedData = Schemas["TurnCompletedData"];
+export type OpenApiTurnFailedData = Schemas["TurnFailedData"];
+export type OpenApiTurnSealedData = Schemas["TurnSealedData"];
+export type OpenApiTurnStartedData = Schemas["TurnStartedData"];
+export type OpenApiUpdateAgentRequest = Schemas["UpdateAgentRequest"];
+export type OpenApiUpdateAppRequest = Schemas["UpdateAppRequest"];
+export type OpenApiUpdateDeclarativeCapabilityRequest =
+  Schemas["UpdateDeclarativeCapabilityRequest"];
+export type OpenApiUpdateFileRequest = Schemas["UpdateFileRequest"];
+export type OpenApiUpdateHarnessRequest = Schemas["UpdateHarnessRequest"];
+export type OpenApiUpdateKnowledgeBaseRequest = Schemas["UpdateKnowledgeBaseRequest"];
+export type OpenApiUpdateKnowledgeEntryRequest = Schemas["UpdateKnowledgeEntryRequest"];
+export type OpenApiUpdateKnowledgeIndexRequest = Schemas["UpdateKnowledgeIndexRequest"];
+export type OpenApiUpdateMcpServerRequest = Schemas["UpdateMcpServerRequest"];
+export type OpenApiUpdateMemoryFileRequest = Schemas["UpdateMemoryFileRequest"];
+export type OpenApiUpdateMemoryRequest = Schemas["UpdateMemoryRequest"];
+export type OpenApiUpdateModelRequest = Schemas["UpdateModelRequest"];
+export type OpenApiUpdateOrgFeatureFlagsRequest = Schemas["UpdateOrgFeatureFlagsRequest"];
+export type OpenApiUpdateOrganizationRequest = Schemas["UpdateOrganizationRequest"];
+export type OpenApiUpdatePaymentAccountRequest = Schemas["UpdatePaymentAccountRequest"];
+export type OpenApiUpdatePaymentPolicyRequest = Schemas["UpdatePaymentPolicyRequest"];
+export type OpenApiUpdateProfileRequest = Schemas["UpdateProfileRequest"];
+export type OpenApiUpdateProviderRequest = Schemas["UpdateProviderRequest"];
+export type OpenApiUpdateSavedReportRequest = Schemas["UpdateSavedReportRequest"];
+export type OpenApiUpdateScheduleRequest = Schemas["UpdateScheduleRequest"];
+export type OpenApiUpdateSessionRequest = Schemas["UpdateSessionRequest"];
+export type OpenApiUpdateSkillRequest = Schemas["UpdateSkillRequest"];
+export type OpenApiUpdateTaskWebhookRequest = Schemas["UpdateTaskWebhookRequest"];
+export type OpenApiUpdateWorkspaceRequest = Schemas["UpdateWorkspaceRequest"];
+export type OpenApiUser = Schemas["User"];
+export type OpenApiValidateSkillRequest = Schemas["ValidateSkillRequest"];
+export type OpenApiVoiceAttachRequest = Schemas["VoiceAttachRequest"];
+export type OpenApiVoiceAttachResponse = Schemas["VoiceAttachResponse"];
+export type OpenApiVoiceCallRequest = Schemas["VoiceCallRequest"];
+export type OpenApiVoiceCallResponse = Schemas["VoiceCallResponse"];
+export type OpenApiVoiceClientSecretRequest = Schemas["VoiceClientSecretRequest"];
+export type OpenApiVoiceClientSecretResponse = Schemas["VoiceClientSecretResponse"];
+export type OpenApiVoiceEndRequest = Schemas["VoiceEndRequest"];
+export type OpenApiVoiceEndResponse = Schemas["VoiceEndResponse"];
+export type OpenApiVoiceSessionEndedData = Schemas["VoiceSessionEndedData"];
+export type OpenApiVoiceSessionFailedData = Schemas["VoiceSessionFailedData"];
+export type OpenApiVoiceSessionOptions = Schemas["VoiceSessionOptions"];
+export type OpenApiVoiceSessionResponse_VoiceCallResponse =
+  Schemas["VoiceSessionResponse_VoiceCallResponse"];
+export type OpenApiVoiceSessionStartedData = Schemas["VoiceSessionStartedData"];
+export type OpenApiVoiceTranscriptData = Schemas["VoiceTranscriptData"];
+export type OpenApiWebhookInvocationResponse = Schemas["WebhookInvocationResponse"];
+export type OpenApiWithUrls_Agent = Schemas["WithUrls_Agent"];
+export type OpenApiWithUrls_App = Schemas["WithUrls_App"];
+export type OpenApiWithUrls_CapabilityInfo = Schemas["WithUrls_CapabilityInfo"];
+export type OpenApiWithUrls_DeclarativeCapability = Schemas["WithUrls_DeclarativeCapability"];
+export type OpenApiWithUrls_Harness = Schemas["WithUrls_Harness"];
+export type OpenApiWithUrls_McpServer = Schemas["WithUrls_McpServer"];
+export type OpenApiWithUrls_Model = Schemas["WithUrls_Model"];
+export type OpenApiWithUrls_ModelWithProvider = Schemas["WithUrls_ModelWithProvider"];
+export type OpenApiWithUrls_Provider = Schemas["WithUrls_Provider"];
+export type OpenApiWithUrls_ResourceWithCounts_Agent = Schemas["WithUrls_ResourceWithCounts_Agent"];
+export type OpenApiWithUrls_ResourceWithCounts_Harness =
+  Schemas["WithUrls_ResourceWithCounts_Harness"];
+export type OpenApiWithUrls_Session = Schemas["WithUrls_Session"];
+export type OpenApiWithUrls_Skill = Schemas["WithUrls_Skill"];
+export type OpenApiWorkerResponse = Schemas["WorkerResponse"];
+export type OpenApiWorkersListResponse = Schemas["WorkersListResponse"];
+export type OpenApiWorkersSummaryResponse = Schemas["WorkersSummaryResponse"];
+export type OpenApiWorkflowEventResponse = Schemas["WorkflowEventResponse"];
+export type OpenApiWorkflowEventsListResponse = Schemas["WorkflowEventsListResponse"];
+export type OpenApiWorkflowResponse = Schemas["WorkflowResponse"];
+export type OpenApiWorkflowsListResponse = Schemas["WorkflowsListResponse"];
+export type OpenApiWorkspaceResponse = Schemas["WorkspaceResponse"];
+export type OpenApiagentId = Schemas["agentId"];
+export type OpenApiagentverId = Schemas["agentverId"];
+export type OpenApieventId = Schemas["eventId"];
+export type OpenApiharnessId = Schemas["harnessId"];
+export type OpenApipayacctId = Schemas["payacctId"];
+export type OpenApipayattId = Schemas["payattId"];
+export type OpenApipaypolId = Schemas["paypolId"];
+
+export type AddA2aChannelHttpRequest = Schemas["AddA2aChannelHttpRequest"];
+export type AddA2aChannelOutput = Schemas["AddA2aChannelOutput"];
+export type AddApiEndpointChannelHttpRequest = Schemas["AddApiEndpointChannelHttpRequest"];
+export type AddApiEndpointChannelOutput = Schemas["AddApiEndpointChannelOutput"];
+export type AgentMessage = Schemas["AgentMessage"];
+export type AllowedAction = Schemas["AllowedAction"];
+export type AppRunListResponse = Schemas["AppRunListResponse"];
+export type BTreeMap = Schemas["BTreeMap"];
+export type BackgroundProgress = Schemas["BackgroundProgress"];
+export type BatchSetSecretsRequest = Schemas["BatchSetSecretsRequest"];
+export type BatchSetSecretsResponse = Schemas["BatchSetSecretsResponse"];
+export type BudgetEventData = Schemas["BudgetEventData"];
+export type CancelStatus = Schemas["CancelStatus"];
+export type CancelTurnResponse = Schemas["CancelTurnResponse"];
+export type CapabilityInfo = Schemas["CapabilityInfo"];
+export type CapabilityLocalizationInfo = Schemas["CapabilityLocalizationInfo"];
+export type CapabilityUsageData = Schemas["CapabilityUsageData"];
+export type CapabilityUsageKind = Schemas["CapabilityUsageKind"];
+export type CapabilityUsageRecord = Schemas["CapabilityUsageRecord"];
+export type CheckAgentNameResponse = Schemas["CheckAgentNameResponse"];
+export type CheckNameResponse = Schemas["CheckNameResponse"];
+export type CircuitBreakerResponse = Schemas["CircuitBreakerResponse"];
+export type CircuitBreakersListResponse = Schemas["CircuitBreakersListResponse"];
+export type ClientToolResult = Schemas["ClientToolResult"];
+export type CommitRequest = Schemas["CommitRequest"];
+export type CommitResult = Schemas["CommitResult"];
+export type CompactionReason = Schemas["CompactionReason"];
+export type CreateBranchRequest = Schemas["CreateBranchRequest"];
+export type CreateDatabaseRequest = Schemas["CreateDatabaseRequest"];
+export type CreateKnowledgeBaseRequest = Schemas["CreateKnowledgeBaseRequest"];
+export type CreateKnowledgeEntryRequest = Schemas["CreateKnowledgeEntryRequest"];
+export type CreateMemoryFileRequest = Schemas["CreateMemoryFileRequest"];
+export type CreateMemorySourceRequest = Schemas["CreateMemorySourceRequest"];
+export type CreateTaskWebhookRequest = Schemas["CreateTaskWebhookRequest"];
+export type CreateWorkspaceRequest = Schemas["CreateWorkspaceRequest"];
+export type DatabaseInfoResponse = Schemas["DatabaseInfoResponse"];
+export type DeferrablePolicy = Schemas["DeferrablePolicy"];
+export type DeleteAccountResponse = Schemas["DeleteAccountResponse"];
+export type DeleteQuery = Schemas["DeleteQuery"];
+export type DeleteResponse = Schemas["DeleteResponse"];
+export type DiffQuery = Schemas["DiffQuery"];
+export type DlqEntryResponse = Schemas["DlqEntryResponse"];
+export type DlqListResponse = Schemas["DlqListResponse"];
+export type EnqueueTaskOptions = Schemas["EnqueueTaskOptions"];
+export type ErrorResponse = Schemas["ErrorResponse"];
+export type EventTypeCountOut = Schemas["EventTypeCountOut"];
+export type EventsSummaryResult = Schemas["EventsSummaryResult"];
+export type ExecutionPhase = Schemas["ExecutionPhase"];
+export type ExportReportQueryRequest = Schemas["ExportReportQueryRequest"];
+export type ExportSavedReportRequest = Schemas["ExportSavedReportRequest"];
+export type ExportUserDataResponse = Schemas["ExportUserDataResponse"];
+export type ExportedOrganization = Schemas["ExportedOrganization"];
+export type ExportedPersonalAccessToken = Schemas["ExportedPersonalAccessToken"];
+export type ExportedUserProfile = Schemas["ExportedUserProfile"];
+export type ExternalActor = Schemas["ExternalActor"];
+export type FeatureFlagMap = Schemas["FeatureFlagMap"];
+export type FieldType = Schemas["FieldType"];
+export type FileWrittenData = Schemas["FileWrittenData"];
+export type Finding = Schemas["Finding"];
+export type ForkSessionRequest = Schemas["ForkSessionRequest"];
+export type FormField = Schemas["FormField"];
+export type GetOrCreateChatSessionRequest = Schemas["GetOrCreateChatSessionRequest"];
+export type GetQuery = Schemas["GetQuery"];
+export type GetResponse = Schemas["GetResponse"];
+export type GetSessionSandboxResponse = Schemas["GetSessionSandboxResponse"];
+export type GitCommitInfo = Schemas["GitCommitInfo"];
+export type GitDiff = Schemas["GitDiff"];
+export type GitDiffEntry = Schemas["GitDiffEntry"];
+export type GitDiffStats = Schemas["GitDiffStats"];
+export type GitHubMemorySourceRequest = Schemas["GitHubMemorySourceRequest"];
+export type GitHubMemorySourceResponse = Schemas["GitHubMemorySourceResponse"];
+export type GitMemorySourceRequest = Schemas["GitMemorySourceRequest"];
+export type GitMemorySourceResponse = Schemas["GitMemorySourceResponse"];
+export type GitRefInfo = Schemas["GitRefInfo"];
+export type GrepResultEntry = Schemas["GrepResultEntry"];
+export type GuardrailAction = Schemas["GuardrailAction"];
+export type GuardrailExample = Schemas["GuardrailExample"];
+export type GuardrailExamplesResponse = Schemas["GuardrailExamplesResponse"];
+export type GuardrailStage = Schemas["GuardrailStage"];
+export type GuardrailsDryRunHit = Schemas["GuardrailsDryRunHit"];
+export type GuardrailsDryRunRequest = Schemas["GuardrailsDryRunRequest"];
+export type GuardrailsDryRunResponse = Schemas["GuardrailsDryRunResponse"];
+export type HarnessPreviewResponse = Schemas["HarnessPreviewResponse"];
+export type ImageContentPart = Schemas["ImageContentPart"];
+export type ImageFileContentPart = Schemas["ImageFileContentPart"];
+export type ImportOkfBundleRequest = Schemas["ImportOkfBundleRequest"];
+export type InputContentPart = Schemas["InputContentPart"];
+export type InputMessage = Schemas["InputMessage"];
+export type KnowledgeBaseResponse = Schemas["KnowledgeBaseResponse"];
+export type KnowledgeEntryResponse = Schemas["KnowledgeEntryResponse"];
+export type KnowledgeIndexDocumentResponse = Schemas["KnowledgeIndexDocumentResponse"];
+export type KnowledgeIndexResponse = Schemas["KnowledgeIndexResponse"];
+export type LeasedResource = Schemas["LeasedResource"];
+export type LeasedResourceStatus = Schemas["LeasedResourceStatus"];
+export type ListAppRunsQuery = Schemas["ListAppRunsQuery"];
+export type ListExecutionsQuery = Schemas["ListExecutionsQuery"];
+export type ListKnowledgeBasesQuery = Schemas["ListKnowledgeBasesQuery"];
+export type ListKnowledgeEntriesQuery = Schemas["ListKnowledgeEntriesQuery"];
+export type ListKnowledgeIndexesQuery = Schemas["ListKnowledgeIndexesQuery"];
+export type ListMemoriesQuery = Schemas["ListMemoriesQuery"];
+export type ListResponse_Agent = Schemas["ListResponse_Agent"];
+export type ListResponse_CapabilityInfo = Schemas["ListResponse_CapabilityInfo"];
+export type ListResponse_DatabaseInfoResponse = Schemas["ListResponse_DatabaseInfoResponse"];
+export type ListResponse_Event = Schemas["ListResponse_Event"];
+export type ListResponse_FileInfo = Schemas["ListResponse_FileInfo"];
+export type ListResponse_GrepResult = Schemas["ListResponse_GrepResult"];
+export type ListResponse_GrepResultEntry = Schemas["ListResponse_GrepResultEntry"];
+export type ListResponse_Harness = Schemas["ListResponse_Harness"];
+export type ListResponse_KeyValueInfo = Schemas["ListResponse_KeyValueInfo"];
+export type ListResponse_KnowledgeBaseResponse = Schemas["ListResponse_KnowledgeBaseResponse"];
+export type ListResponse_KnowledgeEntryResponse = Schemas["ListResponse_KnowledgeEntryResponse"];
+export type ListResponse_KnowledgeIndexDocumentResponse =
+  Schemas["ListResponse_KnowledgeIndexDocumentResponse"];
+export type ListResponse_KnowledgeIndexResponse = Schemas["ListResponse_KnowledgeIndexResponse"];
+export type ListResponse_McpServer = Schemas["ListResponse_McpServer"];
+export type ListResponse_MemoryFileInfo = Schemas["ListResponse_MemoryFileInfo"];
+export type ListResponse_MemoryResponse = Schemas["ListResponse_MemoryResponse"];
+export type ListResponse_Message = Schemas["ListResponse_Message"];
+export type ListResponse_OrganizationResponse = Schemas["ListResponse_OrganizationResponse"];
+export type ListResponse_SavedReport = Schemas["ListResponse_SavedReport"];
+export type ListResponse_SecretInfo = Schemas["ListResponse_SecretInfo"];
+export type ListResponse_Skill = Schemas["ListResponse_Skill"];
+export type ListResponse_TaskWebhookResponse = Schemas["ListResponse_TaskWebhookResponse"];
+export type ListResponse_User = Schemas["ListResponse_User"];
+export type ListResponse_WithUrls_App = Schemas["ListResponse_WithUrls_App"];
+export type ListResponse_WithUrls_DeclarativeCapability =
+  Schemas["ListResponse_WithUrls_DeclarativeCapability"];
+export type ListResponse_WithUrls_McpServer = Schemas["ListResponse_WithUrls_McpServer"];
+export type ListResponse_WithUrls_Model = Schemas["ListResponse_WithUrls_Model"];
+export type ListResponse_WithUrls_ModelWithProvider =
+  Schemas["ListResponse_WithUrls_ModelWithProvider"];
+export type ListResponse_WithUrls_Provider = Schemas["ListResponse_WithUrls_Provider"];
+export type ListResponse_WithUrls_ResourceWithCounts_Harness =
+  Schemas["ListResponse_WithUrls_ResourceWithCounts_Harness"];
+export type ListResponse_WithUrls_Skill = Schemas["ListResponse_WithUrls_Skill"];
+export type ListResponse_WorkspaceResponse = Schemas["ListResponse_WorkspaceResponse"];
+export type ListSchedulesQuery = Schemas["ListSchedulesQuery"];
+export type ListWorkspacesQuery = Schemas["ListWorkspacesQuery"];
+export type LlmPromptCacheInfo = Schemas["LlmPromptCacheInfo"];
+export type LlmRequestOptions = Schemas["LlmRequestOptions"];
+export type LlmRetryInfo = Schemas["LlmRetryInfo"];
+export type LlmToolSearchInfo = Schemas["LlmToolSearchInfo"];
+export type LogQuery = Schemas["LogQuery"];
+export type ManageSessionSandboxRequest = Schemas["ManageSessionSandboxRequest"];
+export type ManageSessionSandboxResponse = Schemas["ManageSessionSandboxResponse"];
+export type ManualMemorySourceResponse = Schemas["ManualMemorySourceResponse"];
+export type McpErrorCategory = Schemas["McpErrorCategory"];
+export type McpErrorCode = Schemas["McpErrorCode"];
+export type McpExecuteError = Schemas["McpExecuteError"];
+export type McpToolAnnotations = Schemas["McpToolAnnotations"];
+export type MemoryFile = Schemas["MemoryFile"];
+export type MemoryFileInfo = Schemas["MemoryFileInfo"];
+export type MemoryResponse = Schemas["MemoryResponse"];
+export type MemorySourceResponse = Schemas["MemorySourceResponse"];
+export type MessageBody = Schemas["MessageBody"];
+export type ModelSource = Schemas["ModelSource"];
+export type OkfFileInput = Schemas["OkfFileInput"];
+export type OkfImportSummary = Schemas["OkfImportSummary"];
+export type OrganizationResponse = Schemas["OrganizationResponse"];
+export type OutputMessageReplacedData = Schemas["OutputMessageReplacedData"];
+export type PaginatedResponse_Session = Schemas["PaginatedResponse_Session"];
+export type PaginatedResponse_WithUrls_CapabilityInfo =
+  Schemas["PaginatedResponse_WithUrls_CapabilityInfo"];
+export type PaginatedResponse_WithUrls_ResourceWithCounts_Agent =
+  Schemas["PaginatedResponse_WithUrls_ResourceWithCounts_Agent"];
+export type PaginatedResponse_WithUrls_Session = Schemas["PaginatedResponse_WithUrls_Session"];
+export type PostTaskMessageBody = Schemas["PostTaskMessageBody"];
+export type ProjectorRunQuery = Schemas["ProjectorRunQuery"];
+export type PromptCacheStrategy = Schemas["PromptCacheStrategy"];
+export type ReasonRecoveredData = Schemas["ReasonRecoveredData"];
+export type RecoveryMode = Schemas["RecoveryMode"];
+export type RegenerateA2aApiKeyOutput = Schemas["RegenerateA2aApiKeyOutput"];
+export type RegenerateApiEndpointApiKeyOutput = Schemas["RegenerateApiEndpointApiKeyOutput"];
+export type ReportingBackfillRequest = Schemas["ReportingBackfillRequest"];
+export type ResolveOrgResponse = Schemas["ResolveOrgResponse"];
+export type ResourceStatsResponse = Schemas["ResourceStatsResponse"];
+export type RiskLevel = Schemas["RiskLevel"];
+export type ScheduleExecutionResponse = Schemas["ScheduleExecutionResponse"];
+export type ScheduleExecutionsListResponse = Schemas["ScheduleExecutionsListResponse"];
+export type ScheduleResponse = Schemas["ScheduleResponse"];
+export type ScheduleStatsResponse = Schemas["ScheduleStatsResponse"];
+export type ScheduleTargetResponse = Schemas["ScheduleTargetResponse"];
+export type SchedulesListResponse = Schemas["SchedulesListResponse"];
+export type SchemaResponse = Schemas["SchemaResponse"];
+export type SendSignalRequest = Schemas["SendSignalRequest"];
+export type SessionRef = Schemas["SessionRef"];
+export type SessionSandboxAction = Schemas["SessionSandboxAction"];
+export type SessionSandboxStatusValue = Schemas["SessionSandboxStatusValue"];
+export type SessionStatsResponse = Schemas["SessionStatsResponse"];
+export type SessionTaskEventData = Schemas["SessionTaskEventData"];
+export type SideEffectClass = Schemas["SideEffectClass"];
+export type StatRequest = Schemas["StatRequest"];
+export type SubmitToolResultsRequest = Schemas["SubmitToolResultsRequest"];
+export type SubmitToolResultsResponse = Schemas["SubmitToolResultsResponse"];
+export type SuccessResponse = Schemas["SuccessResponse"];
+export type SwitchOrgRequest = Schemas["SwitchOrgRequest"];
+export type SwitchOrgResponse = Schemas["SwitchOrgResponse"];
+export type TaskMessageDirection = Schemas["TaskMessageDirection"];
+export type TaskMessageEventData = Schemas["TaskMessageEventData"];
+export type TaskResponse = Schemas["TaskResponse"];
+export type TaskWebhookResponse = Schemas["TaskWebhookResponse"];
+export type TasksListResponse = Schemas["TasksListResponse"];
+export type TextContentPart = Schemas["TextContentPart"];
+export type ToolCallContentPart = Schemas["ToolCallContentPart"];
+export type ToolCallRepairedData = Schemas["ToolCallRepairedData"];
+export type ToolResultContentPart = Schemas["ToolResultContentPart"];
+export type TranscriptRepairAction = Schemas["TranscriptRepairAction"];
+export type TranscriptRepairedData = Schemas["TranscriptRepairedData"];
+export type TurnCancelledData = Schemas["TurnCancelledData"];
+export type TurnSealedData = Schemas["TurnSealedData"];
+export type UpdateKnowledgeBaseRequest = Schemas["UpdateKnowledgeBaseRequest"];
+export type UpdateKnowledgeEntryRequest = Schemas["UpdateKnowledgeEntryRequest"];
+export type UpdateMemoryFileRequest = Schemas["UpdateMemoryFileRequest"];
+export type UpdateOrgFeatureFlagsRequest = Schemas["UpdateOrgFeatureFlagsRequest"];
+export type UpdateSavedReportRequest = Schemas["UpdateSavedReportRequest"];
+export type UpdateTaskWebhookRequest = Schemas["UpdateTaskWebhookRequest"];
+export type UpdateWorkspaceRequest = Schemas["UpdateWorkspaceRequest"];
+export type VoiceAttachRequest = Schemas["VoiceAttachRequest"];
+export type VoiceAttachResponse = Schemas["VoiceAttachResponse"];
+export type VoiceCallRequest = Schemas["VoiceCallRequest"];
+export type VoiceCallResponse = Schemas["VoiceCallResponse"];
+export type VoiceClientSecretRequest = Schemas["VoiceClientSecretRequest"];
+export type VoiceClientSecretResponse = Schemas["VoiceClientSecretResponse"];
+export type VoiceEndRequest = Schemas["VoiceEndRequest"];
+export type VoiceEndResponse = Schemas["VoiceEndResponse"];
+export type VoiceSessionEndedData = Schemas["VoiceSessionEndedData"];
+export type VoiceSessionFailedData = Schemas["VoiceSessionFailedData"];
+export type VoiceSessionOptions = Schemas["VoiceSessionOptions"];
+export type VoiceSessionResponse_VoiceCallResponse =
+  Schemas["VoiceSessionResponse_VoiceCallResponse"];
+export type VoiceSessionStartedData = Schemas["VoiceSessionStartedData"];
+export type VoiceTranscriptData = Schemas["VoiceTranscriptData"];
+export type WebhookInvocationResponse = Schemas["WebhookInvocationResponse"];
+export type WithUrls_Agent = Schemas["WithUrls_Agent"];
+export type WithUrls_App = Schemas["WithUrls_App"];
+export type WithUrls_CapabilityInfo = Schemas["WithUrls_CapabilityInfo"];
+export type WithUrls_DeclarativeCapability = Schemas["WithUrls_DeclarativeCapability"];
+export type WithUrls_Harness = Schemas["WithUrls_Harness"];
+export type WithUrls_McpServer = Schemas["WithUrls_McpServer"];
+export type WithUrls_Model = Schemas["WithUrls_Model"];
+export type WithUrls_ModelWithProvider = Schemas["WithUrls_ModelWithProvider"];
+export type WithUrls_Provider = Schemas["WithUrls_Provider"];
+export type WithUrls_ResourceWithCounts_Agent = Schemas["WithUrls_ResourceWithCounts_Agent"];
+export type WithUrls_ResourceWithCounts_Harness = Schemas["WithUrls_ResourceWithCounts_Harness"];
+export type WithUrls_Session = Schemas["WithUrls_Session"];
+export type WithUrls_Skill = Schemas["WithUrls_Skill"];
+export type WorkerResponse = Schemas["WorkerResponse"];
+export type WorkersListResponse = Schemas["WorkersListResponse"];
+export type WorkersSummaryResponse = Schemas["WorkersSummaryResponse"];
+export type WorkflowEventResponse = Schemas["WorkflowEventResponse"];
+export type WorkflowEventsListResponse = Schemas["WorkflowEventsListResponse"];
+export type WorkflowResponse = Schemas["WorkflowResponse"];
+export type WorkflowsListResponse = Schemas["WorkflowsListResponse"];
+export type WorkspaceResponse = Schemas["WorkspaceResponse"];
+export type agentId = Schemas["agentId"];
+export type agentverId = Schemas["agentverId"];
+export type eventId = Schemas["eventId"];
+export type harnessId = Schemas["harnessId"];
+export type payacctId = Schemas["payacctId"];
+export type payattId = Schemas["payattId"];
+export type paypolId = Schemas["paypolId"];
