+++ apps/ui/src/lib/api/types.ts
@@ -1,3 +1,2 @@
-// Barrel re-export: all API types split into domain-specific files under ./types/
-// Existing imports from "@/lib/api/types" continue to work unchanged.
-export * from "./types/index";
+export * from "./schema-types";
+export * from "./legacy-api-types";
