--- apps/ui/src/lib/api/types/index.ts
@@ -1,26 +1,0 @@
-// Barrel re-export for all API types
-// Each domain has its own file; this index re-exports everything
-// so that `import { ... } from "@/lib/api/types"` continues to work.
-
-export * from "./common-types";
-export * from "./agent-types";
-export * from "./agent-identity-types";
-export * from "./session-types";
-export * from "./message-types";
-export * from "./event-types";
-export * from "./auth-types";
-export * from "./provider-types";
-export * from "./capability-types";
-export * from "./durable-types";
-export * from "./mcp-types";
-export * from "./skill-types";
-export * from "./image-types";
-export * from "./app-types";
-export * from "./command-types";
-export * from "./eval-types";
-export * from "./budget-types";
-export * from "./payment-types";
-export * from "./memory-types";
-export * from "./knowledge-index-types";
-export * from "./reporting-types";
-export * from "./plugin-types";
