--- apps/ui/src/lib/api/types/provider-types.ts
@@ -1,275 +1,0 @@
-// LLM Provider and Model types
-
-// ============================================
-// LLM Provider types
-// ============================================
-
-export type DriverId =
-  | "openai"
-  | "openrouter"
-  | "azure_openai"
-  | "openai_completions"
-  | "anthropic"
-  | "gemini"
-  | "bedrock"
-  | "mai"
-  | "fireworks";
-
-export type ProviderStatus = "active" | "disabled";
-
-/** Vendor/brand of a model, derived from the backend model registry. */
-export type ModelVendor =
-  | "openai"
-  | "anthropic"
-  | "google"
-  | "nvidia"
-  | "qwen"
-  | "microsoft"
-  | "minimax"
-  | "moonshot"
-  | "xai"
-  | "llmsim";
-
-/**
- * Configuration for linking from the chat UI to a provider's observability
- * dashboard ("trace"/"logs"). URL templates support the `{response_id}`,
- * `{session_id}`, `{turn_id}` and `{model}` placeholders. The backend resolves
- * driver defaults overlaid with the org's stored overrides; `enabled` gates
- * whether links are shown at all.
- */
-export interface ProviderTraceConfig {
-  enabled: boolean;
-  generation_url_template?: string;
-  session_url_template?: string;
-}
-
-export interface Provider {
-  id: string;
-  name: string;
-  provider_type: DriverId;
-  base_url?: string;
-  api_key_set: boolean;
-  status: ProviderStatus;
-  created_at: string;
-  updated_at: string;
-  /** Resolved trace/observability link configuration (driver defaults + overrides). */
-  trace?: ProviderTraceConfig;
-}
-
-export interface Model {
-  id: string;
-  provider_id: string;
-  model_id: string;
-  display_name: string;
-  capabilities: string[];
-  enabled: boolean;
-  is_favorite: boolean;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface ModelWithProvider extends Model {
-  provider_name: string;
-  provider_type: DriverId;
-  /**
-   * Derived: model is configured and ready for use. Currently mirrors the
-   * joined provider's `api_key_set && status === "active"`; not persisted.
-   */
-  healthy: boolean;
-  /** Readonly profile with model capabilities (not persisted to database) */
-  profile?: ModelProfile;
-  /** Vendor/brand from the model registry; drives branding. Not persisted. */
-  model_vendor?: ModelVendor;
-}
-
-// ============================================
-// LLM Model Profile types
-// Based on models.dev structure
-// ============================================
-
-/** Cost information for the model (per million tokens in USD) */
-export interface ModelCost {
-  /** Input cost per million tokens */
-  input: number;
-  /** Output cost per million tokens */
-  output: number;
-  /** Cached read cost per million tokens, if supported */
-  cache_read?: number;
-  /** Tiered pricing above certain context thresholds */
-  cost_tiers?: CostTier[];
-}
-
-/** A pricing tier that activates above a context token threshold */
-export interface CostTier {
-  /** Context token threshold above which this tier applies */
-  above_tokens: number;
-  /** Input cost per million tokens (USD) for this tier */
-  input: number;
-  /** Output cost per million tokens (USD) for this tier */
-  output: number;
-  /** Cached read cost per million tokens (USD) for this tier, if supported */
-  cache_read?: number;
-}
-
-/** Token limits for the model */
-export interface ModelLimits {
-  /** Maximum context window size in tokens */
-  context: number;
-  /** Maximum input tokens (if different from context - output) */
-  input?: number;
-  /** Maximum output tokens */
-  output: number;
-  /** Maximum images or PDF pages per request */
-  max_media?: number;
-}
-
-/** Modality type */
-export type Modality = "text" | "image" | "audio" | "video" | "pdf";
-
-/** Model modalities for input and output */
-export interface ModelModalities {
-  /** Supported input modalities */
-  input: Modality[];
-  /** Supported output modalities */
-  output: Modality[];
-}
-
-/** Reasoning effort level for models that support it */
-export type ReasoningEffort = "none" | "minimal" | "low" | "medium" | "high" | "xhigh";
-
-/** Named reasoning effort value for UI display */
-export interface ReasoningEffortValue {
-  /** The API value (e.g., "low", "medium") */
-  value: ReasoningEffort;
-  /** Display name (e.g., "Low", "Medium") */
-  name: string;
-}
-
-/** Reasoning effort configuration for a model */
-export interface ReasoningEffortConfig {
-  /** Available reasoning effort values for this model */
-  values: ReasoningEffortValue[];
-  /** Default reasoning effort for this model */
-  default: ReasoningEffort;
-}
-
-/**
- * LLM Model Profile describing model capabilities
- * Based on models.dev structure (https://models.dev/api.json)
- */
-export interface ModelProfile {
-  /** Display name of the model */
-  name: string;
-  /** Model family (e.g., "gpt-4o", "claude-3-5-sonnet") */
-  family: string;
-  /** Short human-readable description of the model's strengths and intended use */
-  description?: string;
-  /** Release date (YYYY-MM-DD format) */
-  release_date?: string;
-  /** Last updated date (YYYY-MM-DD format) */
-  last_updated?: string;
-  /** Whether the model supports file/image attachments */
-  attachment: boolean;
-  /** Whether the model has reasoning/chain-of-thought capabilities */
-  reasoning: boolean;
-  /** Whether temperature control is supported */
-  temperature: boolean;
-  /** Knowledge cutoff date (YYYY-MM-DD format) */
-  knowledge?: string;
-  /** Whether the model supports tool/function calling */
-  tool_call: boolean;
-  /** Whether the model supports structured output (JSON mode) */
-  structured_output: boolean;
-  /** Whether the model has open weights */
-  open_weights: boolean;
-  /** Cost per million tokens */
-  cost?: ModelCost;
-  /** Token limits */
-  limits?: ModelLimits;
-  /** Supported modalities */
-  modalities?: ModelModalities;
-  /** Reasoning effort configuration (for reasoning models) */
-  reasoning_effort?: ReasoningEffortConfig;
-  /** Provider-advertised request parameters supported by this model */
-  supported_parameters?: string[];
-  /** Whether the model supports tool_search (deferred tool loading) */
-  tool_search?: boolean;
-  /** Whether the model supports native execution phases */
-  supports_phases?: boolean;
-}
-
-export interface CreateProviderRequest {
-  name: string;
-  provider_type: DriverId;
-  base_url?: string;
-  api_key?: string;
-  /** Typed credential fields keyed by the driver's credential-schema field names. */
-  credentials?: Record<string, string>;
-  trace?: ProviderTraceConfig;
-}
-
-export interface UpdateProviderRequest {
-  name?: string;
-  provider_type?: DriverId;
-  base_url?: string;
-  api_key?: string;
-  /** Typed credential fields keyed by the driver's credential-schema field names. */
-  credentials?: Record<string, string>;
-  status?: ProviderStatus;
-  trace?: ProviderTraceConfig;
-}
-
-/** A single declared credential field (mirrors core `FormField`). */
-export interface CredentialFormField {
-  name: string;
-  label: string;
-  field_type: "password" | "text" | "url";
-  required: boolean;
-  placeholder?: string;
-  help_text?: string;
-  /** Default value the form pre-fills (e.g. an OAuth scope or AWS region). */
-  default_value?: string;
-  /** Mutually-exclusive group label; fields sharing it are one credential method. */
-  group?: string;
-}
-
-/** A driver's declared credential schema (mirrors core `CredentialFormSchema`). */
-export interface CredentialFormSchema {
-  fields: CredentialFormField[];
-  instructions_markdown: string;
-}
-
-/** Per-driver credential schema from `GET /v1/providers/config`. */
-export interface DriverCredentialInfo {
-  driver: DriverId;
-  credential_schema: CredentialFormSchema;
-  supports_oauth: boolean;
-}
-
-/** Response shape of `GET /v1/providers/config`. */
-export interface ProvidersConfigResponse {
-  policies: Record<string, boolean>;
-  drivers: DriverCredentialInfo[];
-}
-
-export interface CreateModelRequest {
-  model_id: string;
-  display_name: string;
-  capabilities?: string[];
-  enabled?: boolean;
-  is_favorite?: boolean;
-}
-
-export interface UpdateModelRequest {
-  provider_id?: string;
-  model_id?: string;
-  display_name?: string;
-  capabilities?: string[];
-  enabled?: boolean;
-  is_favorite?: boolean;
-}
-
-// Response from syncing models from a provider
-export type SyncModelsResponse =
-  | { status: "success"; created: number; updated: number; stale: number }
-  | { status: "not_supported" };
